+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#407](https://github.com/green-code-initiative/creedengo-rules-specifications/pull/407): GCI115 - Python - Detect a for loop and suggest a list comprehension
 - [#481](https://github.com/green-code-initiative/creedengo-rules-specifications/pull/481) Add rule GCI114 for Python: Avoid range(len()) pattern, prefer direct iteration or enumerate()
 - Add search words in rules file
-- [#XXX](https://github.com/green-code-initiative/creedengo-rules-specifications/pull/XXX) Add rule GCI116 for Java: Force usage of FetchType LAZY for collections on JPA entities
+- [#487](https://github.com/green-code-initiative/creedengo-rules-specifications/pull/487) Add complete analysis for FetchType issue for Java (sonarqube built-in rule S6904)
 
 ### Changed